# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: books
    description: Bookstore catalog operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List books
      description: Returns all books in the bookstore catalog.
      responses:
        "200":
          description: Book catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookList"
    post:
      operationId: createBook
      tags: [books]
      summary: Add a book
      description: Adds a book to the bookstore catalog.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewBook"
      responses:
        "201":
          description: Book added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Invalid book details
  /books/{id}:
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      description: Deletes a book from the bookstore catalog.
      parameters:
        - name: id
          in: path
          required: true
          description: The book's numeric ID.
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Book deleted
        "404":
          description: Book not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Book:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
        title:
          type: string
        author:
          type: string
      required:
        - id
        - title
        - author
    BookList:
      type: array
      items:
        $ref: "#/components/schemas/Book"
    NewBook:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        author:
          type: string
          minLength: 1
          maxLength: 160
      required:
        - title
        - author